backend/requirements.txt
blinker==1.9.0
click==8.3.1
colorama==0.4.6
Flask==3.1.2
flask-cors==6.0.2
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
psycopg==3.3.2
psycopg-binary==3.3.2
python-dotenv==1.2.1
tzdata==2025.3
Werkzeug==3.1.4
